changed depth image save type to UINT16

+++ irmark1/parts/datastore.py
@@ -236,7 +236,7 @@ class Tub(object):
                 img.save(os.path.join(self.path, name))
                 json_data[key]=name
             elif typ == 'lossless_image_array':
-                img = Image.fromarray(np.uint8(val))
+                img = Image.fromarray(np.uint16(val))
                 name = self.make_file_name(key, ext='.png')
                 img.save(os.path.join(self.path, name), compress_level=0)
                 json_data[key]=name
